[qc] Intergrate --trimmomatic option into qc wrapper command
This changes how many samples are processed in parallel.  There is now a
hard limit (6, subject to future revision) that is relatively low, since
it is estimated that qc is mostly limited by storage bandwith.  CPUs are
distributed among those (<=6) worker threads and a resulting --threads
option passed to qc-sample assuming it gets passed to Trimmomatic even
if the --trimmomatic option is not given.  This anticipates a future
default to use Trimmomatic.