[binning] Drop assumption that sample directories are immediate subdirectories
Previously, the $PWD had to contain the sample directories as immediate
sub-directories because of a limitation in parsing the sample id out of
the coverage file paths in the merge-coverage call.  This patch makes
merge-coverage generate (what hopefully will be) the sample ids from the
variable parts of the coverage file paths.  If omics-binning is invoked
as it was intended before and described above, then nothing should change.
But should the sample directories be sprinkled across several parent
directories then the resulting sample ids in the coverage file used as
input to CONCOCT will also contain parts of those parent directories.