[geomicro env module] Track module file version 2
This complements commit 73f6717b7 which actually added a make rule
installing the geomicro env module to flux, including the comics
executable (which was all that was mentioned in that commit message)

The version 2 module omits all the other stuff, so only makes comics
available.  As a reminder:  it seems that we should never change a
module file in deployment as that confuses the module loading machinery,
always add a new (higher) version of the module file.