[asm] Enable compressed input for idba_ud
This also fixes a long standing issue, that idba actually does not
accepts pipes as input, as it opens and reads the input files twice
The solution here is to create a temporary file under $TMPDIR or /tmp.
So users need to ensure that there is enough disk space.  The tempfile
gets deleted unless idba crashes.