[comics]  Add --norc and --rcfile options with defaults
The default is to run a provided bashrc that is stored in the container.
This should work in current and older versions of singularity and not
depend on hacks of singularies command scripts as more recent versions
of singularity seem to run their own command (shell, exec, etc.) scripts
and not the one stored inside the containers /.singularity.d/action
directories.