[qc] Simplify data flow, remove piping
The attempt to pipe as much data between the tools and reduce saving
data on disk turned out to be too complex, so this patch removes almost
all of it.  Most data will now be created locally on /tmp and copied (as
needed, in the background if possible) to the work directory.