comics: allow specifying container with -i without suffix
This will be clever enough to pick a .sandbox when the --write option is
used.  Also with --write only the locations where sandboxen are expected
are searched.  Without --write the sandbox locations are search last, so
when invoking:

    comics -i sid

we enter the default published sid.img but when running:

    comics -wi sid

then we get sid.sandbox writable.  To enter sid.sandbox read-only, run:

    comics -i sid.sandbox

Also giving the full filename for an image in the current directory will
load that local image and giving the full path to the -i option also
takes precedence over any auto-detection.