omics.shared: use pandas+numpy to store the count array
Uses numpy.stack to build the array row for row, reading in the data
through a generator.  The generator part is triggers a deprecation
warning by numpy.  A DataFrame is then put on top of the array.
Generating the array this way allows for more flexibility in picking
column and checking for consistency and is also faster than using pandas
read_table().  We want the DataFrame over the plain ndarray, as that
makes it easier to keep track of sample and OTU names while manipulating
the array.