shared-get: be very smart when samples names differ between meta and shared files
We now try to do unique prefix matching if the set of names do not
overlap at all or when the option to force this behaviour is given.
This can be helpful when the sample names had the S-number added to make
them unique before running the mothur processing but the meta data still
has the original names from the sequencing core.

In such a case it is also likely that the meta data has been manually
amended to disambiguate the offending sample name resulting in some
exact matches to the shared file.  This case is what the
--force-prefix-matching option is useful and come to think of it, it may
always be needed in praxis.

Dash to underscore substitution is now guarded by a test of usefulness.